# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Crow's Eye Marketing Platform API",
    "description": "REST API for the Crow's Eye marketing automation platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/analytics/": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Get Analytics",
        "description": "Get performance analytics for user's content.\n\n**Tier Required:** Pro+ ($19/month and above)\n\nProvides comprehensive analytics including:\n- Engagement rates and trends\n- Top performing content\n- Platform-specific metrics\n- Growth tracking\n- Audience insights",
        "operationId": "get_analytics_analytics__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "30d",
              "title": "Period"
            }
          },
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Platform"
            }
          },
          {
            "name": "x-user-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/export": {
      "post": {
        "tags": [
          "Analytics"
        ],
        "summary": "Export Analytics",
        "description": "Export analytics data in CSV, JSON, or Excel format.\n\n**Tier Required:** Pro+ ($19/month and above)\n\nExport comprehensive analytics data for external analysis:\n- Custom date ranges\n- Platform filtering\n- Multiple export formats\n- Bulk data export",
        "operationId": "export_analytics_analytics_export_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-user-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/insights": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Get Insights",
        "description": "Get AI-powered insights and recommendations.\n\n**Tier Required:** Pro+ ($19/month and above)",
        "operationId": "get_insights_analytics_insights_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-user-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/competitors": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Get Competitor Analysis",
        "description": "Get competitor analysis and benchmarking.\n\n**Tier Required:** Pro+ ($19/month and above)",
        "operationId": "get_competitor_analysis_analytics_competitors_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-user-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/reports/summary": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Get Summary Report",
        "description": "Get executive summary report.\n\n**Tier Required:** Pro+ ($19/month and above)",
        "operationId": "get_summary_report_analytics_reports_summary_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "30d",
              "title": "Period"
            }
          },
          {
            "name": "x-user-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalyticsResponse": {
        "properties": {
          "period": {
            "type": "string",
            "title": "Period"
          },
          "total_posts": {
            "type": "integer",
            "title": "Total Posts"
          },
          "total_engagement": {
            "type": "integer",
            "title": "Total Engagement"
          },
          "avg_engagement_rate": {
            "type": "number",
            "title": "Avg Engagement Rate"
          },
          "top_performing_content": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Top Performing Content"
          },
          "platform_breakdown": {
            "additionalProperties": true,
            "type": "object",
            "title": "Platform Breakdown"
          },
          "growth_metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Growth Metrics"
          }
        },
        "type": "object",
        "required": [
          "period",
          "total_posts",
          "total_engagement",
          "avg_engagement_rate",
          "top_performing_content",
          "platform_breakdown",
          "growth_metrics"
        ],
        "title": "AnalyticsResponse",
        "description": "Analytics response model."
      },
...
      "ExportRequest": {
        "properties": {
          "start_date": {
            "type": "string",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "title": "End Date"
          },
          "platforms": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platforms"
          },
          "metrics": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics"
          },
          "format": {
            "type": "string",
            "title": "Format",
            "default": "csv"
          }
        },
        "type": "object",
        "required": [
          "start_date",
          "end_date"
        ],
        "title": "ExportRequest",
        "description": "Analytics export request model."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}